% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{queueLength}
\alias{queueLength}
\title{List length of all active queues}
\usage{
queueLength(url = getFlowerURL())
}
\arguments{
\item{url}{Character: Flower URL}
}
\value{
List of active queues
}
\description{
List length of all active queues
}
\examples{
queueLength()
}
