% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskActions.R
\name{taskAbort}
\alias{taskAbort}
\alias{taskRevoke}
\alias{taskTimeout}
\alias{taskRateLimit}
\title{Task operations}
\usage{
taskAbort(id, url = getFlowerURL())

taskRevoke(id, terminate = TRUE, signal = "SIGTERM", url = getFlowerURL())

taskTimeout(
  workername,
  task,
  softLimit = 30,
  hardLimit = 100,
  url = getFlowerURL()
)

taskRateLimit(workername, task, ratelimit, url = getFlowerURL())
}
\arguments{
\item{id}{Character: task identifier}

\item{url}{Character: Flower URL}

\item{terminate}{Boolean: terminate task if running?}

\item{signal}{Character: signal to send when terminating task}

\item{workername}{Character: worker name}

\item{softLimit}{Integer: soft limit}

\item{hardLimit}{Integer: hard limit}

\item{ratelimit}{Integer: rate limit}
}
\value{
Flower response
}
\description{
Task operations
}
\examples{
task <- rownames(taskList())[[1]]
taskAbort(task)
taskRevoke(task)
worker <- names(workerList())[[1]]
taskTimeout(worker, "tasks.add", softLimit=60)
taskRateLimit(worker, "tasks.add", 200)
}
