% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskApply.R
\name{taskApply}
\alias{taskApply}
\alias{taskAsyncApply}
\alias{taskSend}
\title{Execute a task}
\usage{
taskApply(task, args = NULL, kwargs = NULL, url = getFlowerURL())

taskAsyncApply(
  task,
  args = NULL,
  kwargs = NULL,
  options = NULL,
  url = getFlowerURL()
)

taskSend(task, args = NULL, kwargs = NULL, url = getFlowerURL())
}
\arguments{
\item{task}{Character: task to run}

\item{args}{List of arguments}

\item{kwargs}{Dictionary of arguments}

\item{options}{Dictionary of \emph{apply_async} keyword arguments}
}
\value{
Task information:
\itemize{
\item Status
\item Identifier
\item Result (in case of \code{taskApply()} only)
}
}
\description{
Tasks can be executed using:
\itemize{
\item \code{taskApply()}: execute a task and wait for result (better for quick tasks)
\item \code{taskAsyncApply()}: execute a task and immediately return task info without
result (better for long-running tasks)
\item \code{taskSend()}: execute a task (doesn't require task sources)
}
}
\examples{
taskApply("tasks.add")
taskAsyncApply("tasks.add", args=list(a=1, b=2))
taskSend("tasks.add", args=list(a=1, b=2))
}
