% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{taskInfo}
\alias{taskInfo}
\title{Get a task info}
\usage{
taskInfo(id, url = getFlowerURL())
}
\arguments{
\item{id}{Character: task identifier}

\item{url}{Character: Flower URL}
}
\value{
List of active queues
}
\description{
Get a task info
}
\examples{
task <- rownames(taskList())[[1]]
taskInfo(task)
}
