% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{taskList}
\alias{taskList}
\title{List information of tasks}
\usage{
taskList(
  limit = NULL,
  offset = NULL,
  sort_by = c("name", "state", "received", "started"),
  workername = NULL,
  taskname = NULL,
  state = NULL,
  received_start = NULL,
  received_end = NULL,
  table = TRUE,
  url = getFlowerURL()
)
}
\arguments{
\item{limit}{Integer: maximum number of tasks}

\item{offset}{Integer: skip first N tasks}

\item{workername}{Character: filter task by workername}

\item{taskname}{Character: filter tasks by taskname}

\item{state}{Character: filter tasks by state}

\item{table}{Boolean: format response as table? If not, a list is returned}

\item{url}{Character: Flower URL}

\item{sortBy}{Character: sort tasks by attribute (\code{name}, \code{state},
\code{received}, \code{started})}

\item{receivedStart}{Character: filter tasks by received date (must be
greater than) format \%Y-\%m-\%d \%H:\%M}

\item{receivedEnd}{Character: filter tasks by received date (must be less
than) format \%Y-\%m-\%d \%H:\%M}
}
\value{
Data frame or list of task information
}
\description{
List information of tasks
}
\examples{
taskList()
}
