% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{taskTypesList}
\alias{taskTypesList}
\title{List (seen) task types}
\usage{
taskTypesList(url = getFlowerURL())
}
\arguments{
\item{url}{Character: Flower URL}
}
\value{
List of task types
}
\description{
List (seen) task types
}
\examples{
taskTypesList()
}
