% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workerActions.R
\name{workerShutdown}
\alias{workerShutdown}
\alias{workerPoolRestart}
\alias{workerPoolGrow}
\alias{workerPoolShrink}
\alias{workerPoolAutoscale}
\alias{workerQueueAddConsumer}
\alias{workerQueueCancelConsumer}
\title{Worker operations}
\usage{
workerShutdown(workername, url = getFlowerURL())

workerPoolRestart(workername, url = getFlowerURL())

workerPoolGrow(workername, n = 1, url = getFlowerURL())

workerPoolShrink(workername, n = 1, url = getFlowerURL())

workerPoolAutoscale(workername, min, max, url = getFlowerURL())

workerQueueAddConsumer(workername, queue, url = getFlowerURL())

workerQueueCancelConsumer(workername, queue, url = getFlowerURL())
}
\arguments{
\item{workername}{Character: filter info by worker's name}

\item{url}{Character: Flower URL}

\item{n}{Integer: number of pool processes to grow/shrink}

\item{min}{Integer: minimum number of pool processes}

\item{max}{Integer: maximum number of pool processes}

\item{queue}{Character: queue name}
}
\value{
Flower response
}
\description{
Worker operations
}
\examples{
worker <- names(workerList())[[1]]
workerShutdown(worker)
workerPoolRestart(worker)
workerPoolGrow(worker, 4)
workerPoolShrink(worker, 4)
workerPoolAutoscale(worker, 4)
workerQueueAddConsumer(worker, "sample-queue")
workerQueueCancelConsumer(worker, "sample-queue")
}
